+++ ProjectJava/src/main/java/org/example/model/Board.java
@@ -7,34 +7,34 @@ import java.util.ArrayList;
 
 public class Board {
 
-    private final ArrayList<SquareInfo> squares = new ArrayList<>();
-    private BoardType boardType;
+	private final ArrayList<SquareInfo> squares = new ArrayList<>();
+	private BoardType boardType;
 
-    public ArrayList<SquareInfo> getSquares() {
-        return squares;
-    }
+	public ArrayList<SquareInfo> getSquares() {
+		return squares;
+	}
 
-    public int getSquaresSize() {
-        return squares.size() - 1;
-    }
+	public int getSquaresSize() {
+		return squares.size() - 1;
+	}
 
-    public void addSquare(SquareInfo squareToAdd) {
-        squares.add(squareToAdd);
-    }
+	public void addSquare(SquareInfo squareToAdd) {
+		squares.add(squareToAdd);
+	}
 
-    public void setBoardType(BoardType boardType) {
-        this.boardType = boardType;
-    }
+	public void setBoardType(BoardType boardType) {
+		this.boardType = boardType;
+	}
 
-    public boolean isWinner() {
-        return boardType.isWinner();
-    }
+	public boolean isWinner() {
+		return boardType.isWinner();
+	}
 
-    public void getEndAction(int posAfterRoll, Player player) {
-        boardType.endAction(posAfterRoll, getSquaresSize(), player);
-    }
+	public void getEndAction(int posAfterRoll, Player player) {
+		boardType.endAction(posAfterRoll, getSquaresSize(), player);
+	}
 
-    public SquareInfo getApplyAction(int currentPlayerPos) {
-        return squares.get(currentPlayerPos);
-    }
+	public SquareInfo getApplyAction(int currentPlayerPos) {
+		return squares.get(currentPlayerPos);
+	}
 }
