+++ ProjectJava/src/main/java/org/example/model/Player.java
@@ -1,62 +1,62 @@
 package org.example.model;
 
 public class Player {
-    private String name;
-    private int currentPos;
-    private boolean lostTurn;
-    private boolean canPlayAtStart;
-    private boolean hasImmunity;
-    private int round;
+	private String name;
+	private int currentPos;
+	private boolean lostTurn;
+	private boolean canPlayAtStart;
+	private boolean hasImmunity;
+	private int round;
 
-    public String getName() {
-        return name;
-    }
+	public String getName() {
+		return name;
+	}
 
-    public void setName(String name) {
-        this.name = name;
-    }
+	public void setName(String name) {
+		this.name = name;
+	}
 
-    public int getCurrentPos() {
-        return currentPos;
-    }
+	public int getCurrentPos() {
+		return currentPos;
+	}
 
-    public void setCurrentPos(int currentPos) {
-        this.currentPos = currentPos;
-    }
+	public void setCurrentPos(int currentPos) {
+		this.currentPos = currentPos;
+	}
 
-    public boolean isLostTurn() {
-        return lostTurn;
-    }
+	public boolean isLostTurn() {
+		return lostTurn;
+	}
 
-    public void setLostTurn(boolean lostTurn) {
-        this.lostTurn = lostTurn;
-    }
+	public void setLostTurn(boolean lostTurn) {
+		this.lostTurn = lostTurn;
+	}
 
-    public boolean isCanPlayAtStart() {
-        return canPlayAtStart;
-    }
+	public boolean isCanPlayAtStart() {
+		return canPlayAtStart;
+	}
 
-    public void setCanPlayAtStart(boolean canPlayAtStart) {
-        this.canPlayAtStart = canPlayAtStart;
-    }
+	public void setCanPlayAtStart(boolean canPlayAtStart) {
+		this.canPlayAtStart = canPlayAtStart;
+	}
 
-    public boolean isHasImmunity() {
-        return hasImmunity;
-    }
+	public boolean isHasImmunity() {
+		return hasImmunity;
+	}
 
-    public void setHasImmunity(boolean hasImmunity) {
-        this.hasImmunity = hasImmunity;
-    }
+	public void setHasImmunity(boolean hasImmunity) {
+		this.hasImmunity = hasImmunity;
+	}
 
-    public int getRound() {
-        return round;
-    }
+	public int getRound() {
+		return round;
+	}
 
-    public void setRound(int round) {
-        this.round = round;
-    }
+	public void setRound(int round) {
+		this.round = round;
+	}
 
-    public void setRounds() {
-        this.round = round + 1;
-    }
+	public void setRounds() {
+		this.round = round + 1;
+	}
 }
