+++ ProjectJava/src/main/java/org/example/model/boardType/BoardType.java
@@ -3,8 +3,8 @@ package org.example.model.boardType;
 import org.example.model.Player;
 
 public interface BoardType {
-    void endAction(int posAfterRoll, int lastSquare, Player player);
+	void endAction(int posAfterRoll, int lastSquare, Player player);
 
-    boolean isWinner();
+	boolean isWinner();
 }
