+++ ProjectJava/src/main/java/org/example/model/boardType/Classic.java
@@ -3,18 +3,18 @@ package org.example.model.boardType;
 import org.example.model.Player;
 
 public class Classic implements BoardType {
-    boolean winner;
+	boolean winner;
 
-    public void endAction(int posAfterRoll, int lastSquare, Player player) {
-        if (posAfterRoll >= lastSquare) {
-            winner = true;
-            player.setCurrentPos(lastSquare);
-        } else {
-            player.setCurrentPos(posAfterRoll);
-        }
-    }
+	public void endAction(int posAfterRoll, int lastSquare, Player player) {
+		if (posAfterRoll >= lastSquare) {
+			winner = true;
+			player.setCurrentPos(lastSquare);
+		} else {
+			player.setCurrentPos(posAfterRoll);
+		}
+	}
 
-    public boolean isWinner() {
-        return winner;
-    }
+	public boolean isWinner() {
+		return winner;
+	}
 }
