+++ ProjectJava/src/main/java/org/example/model/boardType/Loop.java
@@ -1,32 +1,32 @@
 package org.example.model.boardType;
 
 import org.example.model.Player;
-import org.example.ui.Screen;
+import org.example.model.ui.Screen;
 
 public class Loop implements BoardType {
-    boolean winner;
-    int rounds;
+	boolean winner;
+	int rounds;
 
-    public Loop(int rounds) {
-        this.rounds = rounds;
-    }
+	public Loop(int rounds) {
+		this.rounds = rounds;
+	}
 
-    public void endAction(int posAfterRoll, int lastSquare, Player player) {
+	public void endAction(int posAfterRoll, int lastSquare, Player player) {
 
-        if (posAfterRoll > lastSquare) {
-            posAfterRoll = posAfterRoll - lastSquare;
-            player.setRounds();
-            Screen.displayMessage(player.getName() + " Looped " + player.getRound() + " Times");
-        }
-        player.setCurrentPos(posAfterRoll);
+		if (posAfterRoll > lastSquare) {
+			posAfterRoll = posAfterRoll - lastSquare;
+			player.setRounds();
+			Screen.displayMessage(player.getName() + " Looped " + player.getRound() + " Times");
+		}
+		player.setCurrentPos(posAfterRoll);
 
-        if (player.getRound() == rounds) {
-            player.setCurrentPos(lastSquare);
-            winner = true;
-        }
-    }
+		if (player.getRound() == rounds) {
+			player.setCurrentPos(lastSquare);
+			winner = true;
+		}
+	}
 
-    public boolean isWinner() {
-        return winner;
-    }
+	public boolean isWinner() {
+		return winner;
+	}
 }
