+++ ProjectJava/src/main/java/org/example/model/gameStart/GameStart.java
@@ -3,6 +3,6 @@ package org.example.model.gameStart;
 import org.example.model.Player;
 
 public interface GameStart {
-    void startCondition(Player currentPlayer);
+	void startCondition(Player currentPlayer);
 }
