+++ ProjectJava/src/main/java/org/example/model/gameStart/NoGameStart.java
@@ -3,6 +3,6 @@ package org.example.model.gameStart;
 import org.example.model.Player;
 
 public class NoGameStart implements GameStart {
-    public void startCondition(Player currentPlayer) {
-    }
+	public void startCondition(Player currentPlayer) {
+	}
 }
