+++ ProjectJava/src/main/java/org/example/model/gameStart/RollDiceToEnter.java
@@ -2,25 +2,25 @@ package org.example.model.gameStart;
 
 import org.example.model.Die;
 import org.example.model.Player;
-import org.example.ui.Screen;
+import org.example.model.ui.Screen;
 
 public class RollDiceToEnter implements GameStart {
-    Die die;
+	Die die;
 
-    public RollDiceToEnter(Die die) {
-        this.die = die;
-    }
+	public RollDiceToEnter(Die die) {
+		this.die = die;
+	}
 
-    public void startCondition(Player currentPlayer) {
-        if (!currentPlayer.isCanPlayAtStart()) {
-            currentPlayer.setCanPlayAtStart(die.rollDie() == 6);
-            Screen.displayMessage("Die number is " + die.rollDie());
+	public void startCondition(Player currentPlayer) {
+		if (! currentPlayer.isCanPlayAtStart()) {
+			currentPlayer.setCanPlayAtStart(die.rollDie() == 6);
+			Screen.displayMessage("Die number is " + die.rollDie());
 
-            if (currentPlayer.isCanPlayAtStart()) {
-                Screen.displayMessage("Entered Board");
-            } else {
-                Screen.displayMessage("Number was not 6, Try next round");
-            }
-        }
-    }
+			if (currentPlayer.isCanPlayAtStart()) {
+				Screen.displayMessage("Entered Board");
+			} else {
+				Screen.displayMessage("Number was not 6, Try next round");
+			}
+		}
+	}
 }
