+++ ProjectJava/src/main/java/org/example/model/interaction/Interaction.java
@@ -3,6 +3,6 @@ package org.example.model.interaction;
 import org.example.model.Player;
 
 public interface Interaction {
-    void meetPlayerAction(Player currentPlayer);
+	void meetPlayerAction(Player currentPlayer);
 }
