+++ ProjectJava/src/main/java/org/example/model/interaction/NoInteraction.java
@@ -3,6 +3,6 @@ package org.example.model.interaction;
 import org.example.model.Player;
 
 public class NoInteraction implements Interaction {
-    public void meetPlayerAction(Player currentPlayer) {
-    }
+	public void meetPlayerAction(Player currentPlayer) {
+	}
 }
