+++ ProjectJava/src/main/java/org/example/model/squares/Empty.java
@@ -3,7 +3,7 @@ package org.example.model.squares;
 import org.example.model.Player;
 
 public class Empty implements SquareInfo {
-    public void applyAction(Player currentPlayer) {
-    }
+	public void applyAction(Player currentPlayer) {
+	}
 }
