+++ ProjectJava/src/main/java/org/example/model/squares/GetImmunity.java
@@ -1,11 +1,11 @@
 package org.example.model.squares;
 
 import org.example.model.Player;
-import org.example.ui.Screen;
+import org.example.model.ui.Screen;
 
 public class GetImmunity implements SquareInfo {
-    public void applyAction(Player currentPlayer) {
-        currentPlayer.setHasImmunity(true);
-        Screen.displayMessage("Got immunity");
-    }
+	public void applyAction(Player currentPlayer) {
+		currentPlayer.setHasImmunity(true);
+		Screen.displayMessage("Got immunity");
+	}
 }
