+++ ProjectJava/src/main/java/org/example/model/squares/Ladder.java
@@ -1,17 +1,17 @@
 package org.example.model.squares;
 
 import org.example.model.Player;
-import org.example.ui.Screen;
+import org.example.model.ui.Screen;
 
 public class Ladder implements SquareInfo {
-    private final int goToSquare;
+	private final int goToSquare;
 
-    public Ladder(int goToSquare) {
-        this.goToSquare = goToSquare;
-    }
+	public Ladder(int goToSquare) {
+		this.goToSquare = goToSquare;
+	}
 
-    public void applyAction(Player currentPlayer) {
-        currentPlayer.setCurrentPos(goToSquare);
-        Screen.displayMessage("Pressed ladder ");
-    }
+	public void applyAction(Player currentPlayer) {
+		currentPlayer.setCurrentPos(goToSquare);
+		Screen.displayMessage("Pressed ladder ");
+	}
 }
