+++ ProjectJava/src/main/java/org/example/model/squares/SquareInfo.java
@@ -3,6 +3,6 @@ package org.example.model.squares;
 import org.example.model.Player;
 
 public interface SquareInfo {
-    void applyAction(Player currentPlayer);
+	void applyAction(Player currentPlayer);
 }
