+++ ProjectJava/src/main/java/org/example/model/squares/SquareLoseTurn.java
@@ -1,16 +1,16 @@
 package org.example.model.squares;
 
 import org.example.model.Player;
-import org.example.ui.Screen;
+import org.example.model.ui.Screen;
 
 public class SquareLoseTurn implements SquareInfo {
-    public void applyAction(Player currentPlayer) {
-        if (currentPlayer.isHasImmunity()) {
-            currentPlayer.setHasImmunity(false);
-            Screen.displayMessage("Used immunity");
-        } else {
-            currentPlayer.setLostTurn(true);
-            Screen.displayMessage("Lost turn next round");
-        }
-    }
+	public void applyAction(Player currentPlayer) {
+		if (currentPlayer.isHasImmunity()) {
+			currentPlayer.setHasImmunity(false);
+			Screen.displayMessage("Used immunity");
+		} else {
+			currentPlayer.setLostTurn(true);
+			Screen.displayMessage("Lost turn next round");
+		}
+	}
 }
