+++ ProjectJava/src/main/java/org/example/model/ui/Screen.java
@@ -1,0 +1,14 @@
+package org.example.model.ui;
+
+import java.util.Scanner;
+
+public class Screen {
+	public static void displayMessage(String message) {
+		System.out.println(message);
+	}
+
+	public static String promptString() {
+		Scanner input = new Scanner(System.in);
+		return input.nextLine();
+	}
+}
