+++ SnakeAndFussyBoardGame.iml
@@ -8,6 +8,6 @@
     </content>
     <orderEntry type="inheritedJdk" />
     <orderEntry type="sourceFolder" forTests="false" />
-    <orderEntry type="library" name="json-20201115" level="application" />
+    <orderEntry type="library" exported="" name="json-20201115" level="application" />
   </component>
 </module>
